Update golden files
