+++ generator/integration_tests/golden/CMakeLists.txt
@@ -80,6 +80,8 @@ add_library(
     golden_thing_admin_options.h
     internal/golden_kitchen_sink_auth_decorator.cc
     internal/golden_kitchen_sink_auth_decorator.h
+    internal/golden_kitchen_sink_connection_impl.cc
+    internal/golden_kitchen_sink_connection_impl.h
     internal/golden_kitchen_sink_logging_decorator.cc
     internal/golden_kitchen_sink_logging_decorator.h
     internal/golden_kitchen_sink_metadata_decorator.cc
@@ -92,6 +94,8 @@ add_library(
     internal/golden_kitchen_sink_stub_factory.h
     internal/golden_thing_admin_auth_decorator.cc
     internal/golden_thing_admin_auth_decorator.h
+    internal/golden_thing_admin_connection_impl.cc
+    internal/golden_thing_admin_connection_impl.h
     internal/golden_thing_admin_logging_decorator.cc
     internal/golden_thing_admin_logging_decorator.h
     internal/golden_thing_admin_metadata_decorator.cc
