+++ generator/integration_tests/golden/golden_thing_admin_connection.cc
@@ -18,6 +18,7 @@
 
 #include "generator/integration_tests/golden/golden_thing_admin_connection.h"
 #include "generator/integration_tests/golden/golden_thing_admin_options.h"
+#include "generator/integration_tests/golden/internal/golden_thing_admin_connection_impl.h"
 #include "generator/integration_tests/golden/internal/golden_thing_admin_option_defaults.h"
 #include "generator/integration_tests/golden/internal/golden_thing_admin_stub_factory.h"
 #include "google/cloud/background_threads.h"
@@ -25,7 +26,6 @@
 #include "google/cloud/grpc_options.h"
 #include "google/cloud/internal/async_long_running_operation.h"
 #include "google/cloud/internal/pagination_range.h"
-#include "google/cloud/internal/retry_loop.h"
 #include <memory>
 
 namespace google {
@@ -196,466 +196,6 @@ GoldenThingAdminConnection::AsyncDropDatabase(
       Status(StatusCode::kUnimplemented, "not implemented"));
 }
 
-namespace {
-class GoldenThingAdminConnectionImpl : public GoldenThingAdminConnection {
- public:
-  GoldenThingAdminConnectionImpl(
-      std::unique_ptr<google::cloud::BackgroundThreads> background,
-      std::shared_ptr<golden_internal::GoldenThingAdminStub> stub,
-      Options const& options)
-      : background_(std::move(background)), stub_(std::move(stub)),
-        retry_policy_prototype_(options.get<GoldenThingAdminRetryPolicyOption>()->clone()),
-        backoff_policy_prototype_(options.get<GoldenThingAdminBackoffPolicyOption>()->clone()),
-        polling_policy_prototype_(options.get<GoldenThingAdminPollingPolicyOption>()->clone()),
-        idempotency_policy_(options.get<GoldenThingAdminConnectionIdempotencyPolicyOption>()->clone()) {}
-
-  ~GoldenThingAdminConnectionImpl() override = default;
-
-  StreamRange<google::test::admin::database::v1::Database> ListDatabases(
-      google::test::admin::database::v1::ListDatabasesRequest request) override {
-    request.clear_page_token();
-    auto stub = stub_;
-    auto retry =
-        std::shared_ptr<GoldenThingAdminRetryPolicy const>(retry_policy());
-    auto backoff = std::shared_ptr<BackoffPolicy const>(
-        backoff_policy());
-    auto idempotency = idempotency_policy()->ListDatabases(request);
-    char const* function_name = __func__;
-    return google::cloud::internal::MakePaginationRange<StreamRange<
-        google::test::admin::database::v1::Database>>(
-        std::move(request),
-        [stub, retry, backoff, idempotency, function_name]
-          (google::test::admin::database::v1::ListDatabasesRequest const& r) {
-          return google::cloud::internal::RetryLoop(
-              retry->clone(), backoff->clone(), idempotency,
-              [stub](grpc::ClientContext& context,
-                     google::test::admin::database::v1::ListDatabasesRequest const& request) {
-                return stub->ListDatabases(context, request);
-              },
-              r, function_name);
-        },
-        [](google::test::admin::database::v1::ListDatabasesResponse r) {
-          std::vector<google::test::admin::database::v1::Database> result(r.databases().size());
-          auto& messages = *r.mutable_databases();
-          std::move(messages.begin(), messages.end(), result.begin());
-          return result;
-        });
-  }
-
-  future<StatusOr<google::test::admin::database::v1::Database>>
-  CreateDatabase(google::test::admin::database::v1::CreateDatabaseRequest const& request) override {
-    auto stub = stub_;
-    return google::cloud::internal::AsyncLongRunningOperation<google::test::admin::database::v1::Database>(
-          background_->cq(), request,
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::test::admin::database::v1::CreateDatabaseRequest const& request) {
-            return stub->AsyncCreateDatabase(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::GetOperationRequest const& request) {
-            return stub->AsyncGetOperation(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::CancelOperationRequest const& request) {
-            return stub->AsyncCancelOperation(cq, std::move(context), request);
-          },
-          &google::cloud::internal::ExtractLongRunningResultResponse<google::test::admin::database::v1::Database>,
-          retry_policy(), backoff_policy(),
-          idempotency_policy()->CreateDatabase(request),
-          polling_policy(), __func__);
-  }
-
-  StatusOr<google::test::admin::database::v1::Database>
-  GetDatabase(
-      google::test::admin::database::v1::GetDatabaseRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->GetDatabase(request),
-        [this](grpc::ClientContext& context,
-            google::test::admin::database::v1::GetDatabaseRequest const& request) {
-          return stub_->GetDatabase(context, request);
-        },
-        request, __func__);
-  }
-
-  future<StatusOr<google::test::admin::database::v1::UpdateDatabaseDdlMetadata>>
-  UpdateDatabaseDdl(google::test::admin::database::v1::UpdateDatabaseDdlRequest const& request) override {
-    auto stub = stub_;
-    return google::cloud::internal::AsyncLongRunningOperation<google::test::admin::database::v1::UpdateDatabaseDdlMetadata>(
-          background_->cq(), request,
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::test::admin::database::v1::UpdateDatabaseDdlRequest const& request) {
-            return stub->AsyncUpdateDatabaseDdl(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::GetOperationRequest const& request) {
-            return stub->AsyncGetOperation(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::CancelOperationRequest const& request) {
-            return stub->AsyncCancelOperation(cq, std::move(context), request);
-          },
-          &google::cloud::internal::ExtractLongRunningResultMetadata<google::test::admin::database::v1::UpdateDatabaseDdlMetadata>,
-          retry_policy(), backoff_policy(),
-          idempotency_policy()->UpdateDatabaseDdl(request),
-          polling_policy(), __func__);
-  }
-
-  Status
-  DropDatabase(
-      google::test::admin::database::v1::DropDatabaseRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->DropDatabase(request),
-        [this](grpc::ClientContext& context,
-            google::test::admin::database::v1::DropDatabaseRequest const& request) {
-          return stub_->DropDatabase(context, request);
-        },
-        request, __func__);
-  }
-
-  StatusOr<google::test::admin::database::v1::GetDatabaseDdlResponse>
-  GetDatabaseDdl(
-      google::test::admin::database::v1::GetDatabaseDdlRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->GetDatabaseDdl(request),
-        [this](grpc::ClientContext& context,
-            google::test::admin::database::v1::GetDatabaseDdlRequest const& request) {
-          return stub_->GetDatabaseDdl(context, request);
-        },
-        request, __func__);
-  }
-
-  StatusOr<google::iam::v1::Policy>
-  SetIamPolicy(
-      google::iam::v1::SetIamPolicyRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->SetIamPolicy(request),
-        [this](grpc::ClientContext& context,
-            google::iam::v1::SetIamPolicyRequest const& request) {
-          return stub_->SetIamPolicy(context, request);
-        },
-        request, __func__);
-  }
-
-  StatusOr<google::iam::v1::Policy>
-  GetIamPolicy(
-      google::iam::v1::GetIamPolicyRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->GetIamPolicy(request),
-        [this](grpc::ClientContext& context,
-            google::iam::v1::GetIamPolicyRequest const& request) {
-          return stub_->GetIamPolicy(context, request);
-        },
-        request, __func__);
-  }
-
-  StatusOr<google::iam::v1::TestIamPermissionsResponse>
-  TestIamPermissions(
-      google::iam::v1::TestIamPermissionsRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->TestIamPermissions(request),
-        [this](grpc::ClientContext& context,
-            google::iam::v1::TestIamPermissionsRequest const& request) {
-          return stub_->TestIamPermissions(context, request);
-        },
-        request, __func__);
-  }
-
-  future<StatusOr<google::test::admin::database::v1::Backup>>
-  CreateBackup(google::test::admin::database::v1::CreateBackupRequest const& request) override {
-    auto stub = stub_;
-    return google::cloud::internal::AsyncLongRunningOperation<google::test::admin::database::v1::Backup>(
-          background_->cq(), request,
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::test::admin::database::v1::CreateBackupRequest const& request) {
-            return stub->AsyncCreateBackup(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::GetOperationRequest const& request) {
-            return stub->AsyncGetOperation(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::CancelOperationRequest const& request) {
-            return stub->AsyncCancelOperation(cq, std::move(context), request);
-          },
-          &google::cloud::internal::ExtractLongRunningResultResponse<google::test::admin::database::v1::Backup>,
-          retry_policy(), backoff_policy(),
-          idempotency_policy()->CreateBackup(request),
-          polling_policy(), __func__);
-  }
-
-  StatusOr<google::test::admin::database::v1::Backup>
-  GetBackup(
-      google::test::admin::database::v1::GetBackupRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->GetBackup(request),
-        [this](grpc::ClientContext& context,
-            google::test::admin::database::v1::GetBackupRequest const& request) {
-          return stub_->GetBackup(context, request);
-        },
-        request, __func__);
-  }
-
-  StatusOr<google::test::admin::database::v1::Backup>
-  UpdateBackup(
-      google::test::admin::database::v1::UpdateBackupRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->UpdateBackup(request),
-        [this](grpc::ClientContext& context,
-            google::test::admin::database::v1::UpdateBackupRequest const& request) {
-          return stub_->UpdateBackup(context, request);
-        },
-        request, __func__);
-  }
-
-  Status
-  DeleteBackup(
-      google::test::admin::database::v1::DeleteBackupRequest const& request) override {
-    return google::cloud::internal::RetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->DeleteBackup(request),
-        [this](grpc::ClientContext& context,
-            google::test::admin::database::v1::DeleteBackupRequest const& request) {
-          return stub_->DeleteBackup(context, request);
-        },
-        request, __func__);
-  }
-
-  StreamRange<google::test::admin::database::v1::Backup> ListBackups(
-      google::test::admin::database::v1::ListBackupsRequest request) override {
-    request.clear_page_token();
-    auto stub = stub_;
-    auto retry =
-        std::shared_ptr<GoldenThingAdminRetryPolicy const>(retry_policy());
-    auto backoff = std::shared_ptr<BackoffPolicy const>(
-        backoff_policy());
-    auto idempotency = idempotency_policy()->ListBackups(request);
-    char const* function_name = __func__;
-    return google::cloud::internal::MakePaginationRange<StreamRange<
-        google::test::admin::database::v1::Backup>>(
-        std::move(request),
-        [stub, retry, backoff, idempotency, function_name]
-          (google::test::admin::database::v1::ListBackupsRequest const& r) {
-          return google::cloud::internal::RetryLoop(
-              retry->clone(), backoff->clone(), idempotency,
-              [stub](grpc::ClientContext& context,
-                     google::test::admin::database::v1::ListBackupsRequest const& request) {
-                return stub->ListBackups(context, request);
-              },
-              r, function_name);
-        },
-        [](google::test::admin::database::v1::ListBackupsResponse r) {
-          std::vector<google::test::admin::database::v1::Backup> result(r.backups().size());
-          auto& messages = *r.mutable_backups();
-          std::move(messages.begin(), messages.end(), result.begin());
-          return result;
-        });
-  }
-
-  future<StatusOr<google::test::admin::database::v1::Database>>
-  RestoreDatabase(google::test::admin::database::v1::RestoreDatabaseRequest const& request) override {
-    auto stub = stub_;
-    return google::cloud::internal::AsyncLongRunningOperation<google::test::admin::database::v1::Database>(
-          background_->cq(), request,
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::test::admin::database::v1::RestoreDatabaseRequest const& request) {
-            return stub->AsyncRestoreDatabase(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::GetOperationRequest const& request) {
-            return stub->AsyncGetOperation(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::CancelOperationRequest const& request) {
-            return stub->AsyncCancelOperation(cq, std::move(context), request);
-          },
-          &google::cloud::internal::ExtractLongRunningResultResponse<google::test::admin::database::v1::Database>,
-          retry_policy(), backoff_policy(),
-          idempotency_policy()->RestoreDatabase(request),
-          polling_policy(), __func__);
-  }
-
-  StreamRange<google::longrunning::Operation> ListDatabaseOperations(
-      google::test::admin::database::v1::ListDatabaseOperationsRequest request) override {
-    request.clear_page_token();
-    auto stub = stub_;
-    auto retry =
-        std::shared_ptr<GoldenThingAdminRetryPolicy const>(retry_policy());
-    auto backoff = std::shared_ptr<BackoffPolicy const>(
-        backoff_policy());
-    auto idempotency = idempotency_policy()->ListDatabaseOperations(request);
-    char const* function_name = __func__;
-    return google::cloud::internal::MakePaginationRange<StreamRange<
-        google::longrunning::Operation>>(
-        std::move(request),
-        [stub, retry, backoff, idempotency, function_name]
-          (google::test::admin::database::v1::ListDatabaseOperationsRequest const& r) {
-          return google::cloud::internal::RetryLoop(
-              retry->clone(), backoff->clone(), idempotency,
-              [stub](grpc::ClientContext& context,
-                     google::test::admin::database::v1::ListDatabaseOperationsRequest const& request) {
-                return stub->ListDatabaseOperations(context, request);
-              },
-              r, function_name);
-        },
-        [](google::test::admin::database::v1::ListDatabaseOperationsResponse r) {
-          std::vector<google::longrunning::Operation> result(r.operations().size());
-          auto& messages = *r.mutable_operations();
-          std::move(messages.begin(), messages.end(), result.begin());
-          return result;
-        });
-  }
-
-  StreamRange<google::longrunning::Operation> ListBackupOperations(
-      google::test::admin::database::v1::ListBackupOperationsRequest request) override {
-    request.clear_page_token();
-    auto stub = stub_;
-    auto retry =
-        std::shared_ptr<GoldenThingAdminRetryPolicy const>(retry_policy());
-    auto backoff = std::shared_ptr<BackoffPolicy const>(
-        backoff_policy());
-    auto idempotency = idempotency_policy()->ListBackupOperations(request);
-    char const* function_name = __func__;
-    return google::cloud::internal::MakePaginationRange<StreamRange<
-        google::longrunning::Operation>>(
-        std::move(request),
-        [stub, retry, backoff, idempotency, function_name]
-          (google::test::admin::database::v1::ListBackupOperationsRequest const& r) {
-          return google::cloud::internal::RetryLoop(
-              retry->clone(), backoff->clone(), idempotency,
-              [stub](grpc::ClientContext& context,
-                     google::test::admin::database::v1::ListBackupOperationsRequest const& request) {
-                return stub->ListBackupOperations(context, request);
-              },
-              r, function_name);
-        },
-        [](google::test::admin::database::v1::ListBackupOperationsResponse r) {
-          std::vector<google::longrunning::Operation> result(r.operations().size());
-          auto& messages = *r.mutable_operations();
-          std::move(messages.begin(), messages.end(), result.begin());
-          return result;
-        });
-  }
-
-  future<StatusOr<google::test::admin::database::v1::Database>>
-  LongRunningWithoutRouting(google::test::admin::database::v1::RestoreDatabaseRequest const& request) override {
-    auto stub = stub_;
-    return google::cloud::internal::AsyncLongRunningOperation<google::test::admin::database::v1::Database>(
-          background_->cq(), request,
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::test::admin::database::v1::RestoreDatabaseRequest const& request) {
-            return stub->AsyncLongRunningWithoutRouting(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::GetOperationRequest const& request) {
-            return stub->AsyncGetOperation(cq, std::move(context), request);
-          },
-          [stub](google::cloud::CompletionQueue& cq,
-                 std::unique_ptr<grpc::ClientContext> context,
-                 google::longrunning::CancelOperationRequest const& request) {
-            return stub->AsyncCancelOperation(cq, std::move(context), request);
-          },
-          &google::cloud::internal::ExtractLongRunningResultResponse<google::test::admin::database::v1::Database>,
-          retry_policy(), backoff_policy(),
-          idempotency_policy()->LongRunningWithoutRouting(request),
-          polling_policy(), __func__);
-  }
-
-  future<StatusOr<google::test::admin::database::v1::Database>>
-  AsyncGetDatabase(
-      google::test::admin::database::v1::GetDatabaseRequest const& request) override {
-    auto& stub = stub_;
-    return google::cloud::internal::AsyncRetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->GetDatabase(request),
-        background_->cq(),
-        [stub](CompletionQueue& cq,
-               std::unique_ptr<grpc::ClientContext> context,
-               google::test::admin::database::v1::GetDatabaseRequest const& request) {
-          return stub->AsyncGetDatabase(cq, std::move(context), request);
-        },
-        request, __func__);
-  }
-
-  future<Status>
-  AsyncDropDatabase(
-      google::test::admin::database::v1::DropDatabaseRequest const& request) override {
-    auto& stub = stub_;
-    return google::cloud::internal::AsyncRetryLoop(
-        retry_policy(), backoff_policy(),
-        idempotency_policy()->DropDatabase(request),
-        background_->cq(),
-        [stub](CompletionQueue& cq,
-               std::unique_ptr<grpc::ClientContext> context,
-               google::test::admin::database::v1::DropDatabaseRequest const& request) {
-          return stub->AsyncDropDatabase(cq, std::move(context), request);
-        },
-        request, __func__);
-  }
-
- private:
-  std::unique_ptr<GoldenThingAdminRetryPolicy> retry_policy() {
-    auto const& options = internal::CurrentOptions();
-    if (options.has<GoldenThingAdminRetryPolicyOption>()) {
-      return options.get<GoldenThingAdminRetryPolicyOption>()->clone();
-    }
-    return retry_policy_prototype_->clone();
-  }
-
-  std::unique_ptr<BackoffPolicy> backoff_policy() {
-    auto const& options = internal::CurrentOptions();
-    if (options.has<GoldenThingAdminBackoffPolicyOption>()) {
-      return options.get<GoldenThingAdminBackoffPolicyOption>()->clone();
-    }
-    return backoff_policy_prototype_->clone();
-  }
-
-  std::unique_ptr<PollingPolicy> polling_policy() {
-    auto const& options = internal::CurrentOptions();
-    if (options.has<GoldenThingAdminPollingPolicyOption>()) {
-      return options.get<GoldenThingAdminPollingPolicyOption>()->clone();
-    }
-    return polling_policy_prototype_->clone();
-  }
-
-  std::unique_ptr<GoldenThingAdminConnectionIdempotencyPolicy> idempotency_policy() {
-    auto const& options = internal::CurrentOptions();
-    if (options.has<GoldenThingAdminConnectionIdempotencyPolicyOption>()) {
-      return options.get<GoldenThingAdminConnectionIdempotencyPolicyOption>()->clone();
-    }
-    return idempotency_policy_->clone();
-  }
-
-  std::unique_ptr<google::cloud::BackgroundThreads> background_;
-  std::shared_ptr<golden_internal::GoldenThingAdminStub> stub_;
-  std::unique_ptr<GoldenThingAdminRetryPolicy const> retry_policy_prototype_;
-  std::unique_ptr<BackoffPolicy const> backoff_policy_prototype_;
-  std::unique_ptr<PollingPolicy const> polling_policy_prototype_;
-  std::unique_ptr<GoldenThingAdminConnectionIdempotencyPolicy> idempotency_policy_;
-};
-}  // namespace
-
 std::shared_ptr<GoldenThingAdminConnection> MakeGoldenThingAdminConnection(
     Options options) {
   internal::CheckExpectedOptions<CommonOptionList, GrpcOptionList,
@@ -665,7 +205,7 @@ std::shared_ptr<GoldenThingAdminConnection> MakeGoldenThingAdminConnection(
   auto background = internal::MakeBackgroundThreadsFactory(options)();
   auto stub = golden_internal::CreateDefaultGoldenThingAdminStub(
     background->cq(), options);
-  return std::make_shared<GoldenThingAdminConnectionImpl>(
+  return std::make_shared<golden_internal::GoldenThingAdminConnectionImpl>(
       std::move(background), std::move(stub), options);
 }
 
@@ -683,7 +223,7 @@ std::shared_ptr<golden::GoldenThingAdminConnection>
 MakeGoldenThingAdminConnection(
     std::shared_ptr<GoldenThingAdminStub> stub, Options options) {
   options = GoldenThingAdminDefaultOptions(std::move(options));
-  return std::make_shared<golden::GoldenThingAdminConnectionImpl>(
+  return std::make_shared<golden_internal::GoldenThingAdminConnectionImpl>(
       internal::MakeBackgroundThreadsFactory(options)(),
       std::move(stub), std::move(options));
 }
